优惠券添加是否限制库存字段

+++ jmp/mp-service/src/main/java/com/vpu/mp/service/pojo/shop/coupon/CouponParam.java
@@ -158,7 +158,7 @@ public class CouponParam {
     /**
      * 是否限制库存 0:限制；1:不限制
      */
-    private Integer limitSurplusFlag;
+    private Byte limitSurplusFlag;
     private Timestamp  createTime;
     private Timestamp  updateTime;
